fix: Ensure wrapping of usage on all subcommands
Previously the command runner was configured to wrap at a certain
line length, but its commands were not.

The Command API is super bizarre though, and unlike CommandRunner where
you can supply a wrapping line length to its constructor, Command
requires that you override the creation of its ArgParser.

So I introduced a base class (MelosCommand) that does this consistently,
and had all of the existing commands inherit from it.

Fixes #70