feat!: Use PUB_HOSTED_URL as pub.dev alternative if defined
The PUB_HOSTED_URL environment variable can be used to override the
URL of the package server used by various dart tools.

I marked this as a breaking change, because it could potentially have
consequences in environments where the variable was defined, but was
previously pulling info from pub.dev.

Fixes #82