feat: add includeCommitId option
Add the option `includeCommitId`, which prints the short commit id in
the changelog without including the link to the commit.

This allows users of git repositories that are not GitHub or GitLab to
have the commit sha present in the changelog instead of just the commit
message.

If both `includeCommitId` and `linkToCommits` are set to true, commit
linking takes presidence.