fix: false positive for non-unique packages
It is possible that `resolvePackages` finds the same package twice, under paths that point to the same location but are represented by `String`s that are not equal.

This change ensures a package is only found once by using functions from  `package:path` for using paths as keys in a `Map`.