fix: run commands in single shell
This change removes `runInShell: true` from `Process.start` when
executing commands. This option is unnecessary since Melos is spawning
a shell itself to execute the command. Using `runInShell: true`
becomes a problem for command chains (e.g. '&&' or '||').
In this case only the first command is executed in the shell that Melos
is spawning and the rest of the commands are executed in the shell
spawned by `Process.start`.

The change also cleans up a few other related bits of code.

Fixes #367