feat: built-in melos command for analyzing projects (#655)
First things first, I'm really excited to work on this issue because this year I made a personal goal to contribute more to open source projects. It makes me happy to be able to do it for a tool that I use every day.

Usually, melos projects use a script to run either dart analyze or flutter analyze, as mentioned by @jpnurmi in issue #424. However, I agree that it would be better to have a built-in command for this. Some newbies to the tool might try running this command and then realize they need extra setup to get the same results.

So, this pull request aims to address this problem by adding the needed functionality. It also includes features like running tasks concurrently, filtering, and support for --fatal-infos and --[no-]fatal-warnings options.

I also took the time to update the readme file and provide the latest help section, as well as update the melos.yaml file and remove the old way to run the analyzer.