feat: format built in command (#657)
This PR introduces a new built-in command called "format" that follows the same implementation logic used in the development of the "analyze" command, which was merged in a previous (PR). The purpose of this change is to eliminate the need to set up a separate script in order to format each package.

For example, instead of using the following YAML configuration:

scripts: format: description: Format Dart code. run: dart format . format:check: description: Check formatting of Dart code. run: dart format --output none --set-exit-if-changed .

The first script can now be replaced with a simpler command: melos format. Similarly, the second script can be replaced with melos format --output none --set-exit-if-changed.

Additionally, this new command supports all melos filtering options and concurrency.