chore: Added melos.yaml json schema (#745)
Melos.yaml schema validation should exist and be maintained together
with `melos.yaml` changes.

The current melos.yaml schema is not updated with the latest options on
the Melos Extension, so things like publish hooks and categories show as
errors and autocomplete.

### Why have it here?

- Easier to maintain, and keep it in sync iwth melos.yaml changes
- In VSCode, you do not need an extension; you can just add it on top of
the `melos.yaml` file.

```yaml
# yaml-language-server: $schema=https://raw.githubusercontent.com/../melos.yaml.schema.json

name: melos_workspace
```

This schema contains changes to [publish
hooks](https://melos.invertase.dev/configuration/overview#categories)
and
[category](https://melos.invertase.dev/configuration/overview#categories).

After this is merged, we can make a PR on the VSCode extension to use
this URL in order to keep it up-to-date.