docs: Update Atsign's project using link (#784)
<!--
  Thanks for contributing!

Provide a description of your changes below and a general summary in the
title

Please look at the following checklist to ensure that your PR can be
accepted quickly:
-->

## Description

- Changed our "Projects Using Melos" link from `at_app` to `at_mono`.

We've since deprecated at_app and archived it. at_mono is a much more
niche / interesting use-case (cross-repo dependency analysis and
testing). The use-cases are not limited to what we are currently doing,
but melos plays a massive role in enabling the capabilities of this
use-case.

We wrote an explainer here:
https://github.com/atsign-foundation/.github/blob/trunk/docs/at_mono.md
(to be completed)
## Type of Change

<!--- Put an `x` in all the boxes that apply: -->

- [ ] ✨ `feat` -- New feature (non-breaking change which adds
functionality)
- [ ] 🛠️ `fix` -- Bug fix (non-breaking change which fixes an issue)
- [ ] ❌ `!` -- Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] 🧹 `refactor` -- Code refactor
- [ ] ✅ `ci` -- Build configuration change
- [x] 📝 `docs` -- Documentation
- [ ] 🗑️ `chore` -- Chore