feat(scripts): allow overriding default scripts with custom ones (#945)
## Description

Fixes https://github.com/invertase/melos/issues/874.

This PR makes it so that if you have a custom script defined in
`pubspec.yaml` that has the same name as a built-in command (e.g.
`format`), that script will be run when you invoke _either_ `melos
<script>` or `melos run <script>`. Before this change, your script would
NOT be run if you ran `melos <script>`, the built-in command would be
run, which is not expected.

**One important behavioral quirk I noticed when making this change:**
Custom scripts are not included in `--help` invocation. This is because:
- An empty workspace is used when `--help` is passed.
- Custom scripts are hidden via `ScriptCommand` having `hidden` set to
`true`

I found this behavior to be unexpected. I was able to make custom
scripts show up in `--help` by changing the above conditions, but I did
_not_ commit that change as I suspect their are historical reasons this
behavior exists (e.g. I noticed that when we _don't_ use an empty
workspace during `--help` invocations, information on `init` shows up
which is weird if you've already run `init`.

## Type of Change

<!--- Put an `x` in all the boxes that apply: -->

- [ ] ✨ `feat` -- New feature (non-breaking change which adds
functionality)
- [ ] 🛠️ `fix` -- Bug fix (non-breaking change which fixes an issue)
- [x] ❌ `!` -- Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] 🧹 `refactor` -- Code refactor
- [ ] ✅ `ci` -- Build configuration change
- [x] 📝 `docs` -- Documentation
- [ ] 🗑️ `chore` -- Chore

---------

Co-authored-by: Brandon Trautmann <8343465+btrautmann@users.noreply.github.com>