feat: add functionality to keep custom intellij modules (#949)
<!--
  Thanks for contributing!

Provide a description of your changes below and a general summary in the
title

Please look at the following checklist to ensure that your PR can be
accepted quickly:
-->

As described in #589 intellij modules are overwritten at each `melos
bootstrap`. The reason for this is that the whole modules.xml file
isrecreated at every bootstrap.

Closes #589 

## Description

- Added xml dependency to read and parse the xml content of the
modules.xml properly
- Added a function to compare the modules inside the current modules.xml
with the melos modules. Then merge them before recreating the
modules.xml.

## Type of Change

<!--- Put an `x` in all the boxes that apply: -->

- [x] ✨ `feat` -- New feature (non-breaking change which adds
functionality)
- [ ] 🛠️ `fix` -- Bug fix (non-breaking change which fixes an issue)
- [ ] ❌ `!` -- Breaking change (fix or feature that would cause existing
functionality to change)
- [ ] 🧹 `refactor` -- Code refactor
- [ ] ✅ `ci` -- Build configuration change
- [ ] 📝 `docs` -- Documentation
- [ ] 🗑️ `chore` -- Chore